Feat/add back button (#1375)
* feat(addBackButton): Removed selector to hide back button

* feat(addBackButton): Wrote logic to toggle back button and refactored to use reusable function

* feat(addBackButton): Renamed file based on new usage

* feat(addBackButton): Formatted code for readability

* feat(addBackButton): Added comment per reviewers request to clarify code

* feat(addBackButton): Renamed file and removed comment per reviewer's request
