--- src/Analysis/GWASResults/Utils/useHideUnneededElements.js
@@ -1,19 +1,0 @@
-import { useEffect } from 'react';
-
-const useHideUnneededElements = () => {
-  const selectorsToHide = [
-    '.analysis-app__description',
-    '.back-link',
-    '.analysis-app__title',
-  ];
-  useEffect(() => {
-    selectorsToHide.forEach((selector) => {
-      document.querySelectorAll(selector).forEach((element) => {
-        const temporaryElement = element;
-        temporaryElement.style.display = 'none';
-      });
-    });
-  }, []);
-};
-
-export default useHideUnneededElements;
